Share cargo registry between app and test containers

+++ test/build.sh
@@ -2,19 +2,26 @@
 
 set -eu
 
+if [[ -z ${TARGET:-""} ]]; then
+    echo "\$TARGET must be specified"
+    exit 1
+fi
+
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 APP_DIR="$SCRIPT_DIR/.."
 cd "$SCRIPT_DIR"
 
 if [[ $TARGET == x86_64-unknown-linux-gnu ]]; then
-    mkdir -p .container/cargo-registry
-    container_image=$(cat "$APP_DIR/building/linux-container-image.txt")
-    podman build -t mullvadvpn-app-tests --build-arg IMAGE="${container_image}" .
+    CARGO_REGISTRY_VOLUME_NAME=${CARGO_REGISTRY_VOLUME_NAME:-"cargo-registry"}
+
+    if ! podman image exists mullvadvpn-app-tests; then
+        container_image=$(cat "$APP_DIR/building/linux-container-image.txt")
+        podman build -t mullvadvpn-app-tests --build-arg IMAGE="${container_image}" .
+    fi
 
     podman run --rm -it \
-        -v "${SCRIPT_DIR}/.container/cargo-registry":/root/.cargo/registry \
-        -v "${APP_DIR}":/src:Z \
-        -e CARGO_HOME=/root/.cargo/registry \
+        -v "${CARGO_REGISTRY_VOLUME_NAME}":/root/.cargo/registry:Z \
+        -v "${APP_DIR}":/src:z \
         -e CARGO_TARGET_DIR=/src/test/target \
         mullvadvpn-app-tests \
         /bin/bash -c "cd /src/test/; cargo build --bin test-runner --bin connection-checker --release --target ${TARGET}"
